linefollow trys

+++ colormeasureTest.py
@@ -27,7 +27,7 @@ parser.add_argument('--lspeed',type=int,default=20,help='rsped!',required=False)
 parser.add_argument('--rspeed',type=int,default=0,help='lspeed!',required=False)
 parser.add_argument('--turn',type=bool,default=False,help='turn!',required=False)
 parser.add_argument('--degrees',type=int,default=990,help='deg!',required=False)
-parser.add_argument('--rot',type=int,default=1,help='rot!',required=False)
+parser.add_argument('--rot',type=float,default=1,help='rot!',required=False)
